Refactor code
Moves the code into a cleaner structure with world, cells and grid in
separate source files.
