+++ exploration/grid-poc/world.py
@@ -4,28 +4,29 @@
 # SPDX-License-Identifier: BSD-2-Clause
 # Copyright © 2024 The Alan Turing Institute
 
-# Example Plants
+# World
 
 from time import sleep
 import numpy as np
 import math
-from enum import Enum
 import copy
 import random
 from math import floor
 import voxels as vxm
 from threading import Thread, Lock
+from cells import (
+    Direction,
+    State,
+    States,
+    Cell,
+    Air,
+    Soil,
+    Rock,
+)
 
-UNSATURATED_PRESSURE_GRADIENT = 0.5
-SATURATED_PRESSURE_GRADIENT = 1.0
-
-class Direction(Enum):
-    LEFT = 0
-    RIGHT = 1
-    BELOW = 2
-    ABOVE = 3
-    FRONT = 4
-    BEHIND = 5
+from plants import (
+    Plant,
+)
 
 def opposite(direction):
     return [
@@ -36,150 +37,6 @@ def opposite(direction):
         Direction.BEHIND,
         Direction.FRONT
     ][direction]
-
-class State():
-    state = {}
-
-class States():
-    state = None
-    incoming = None
-    outgoing = None
-    reproduce = None
-
-    def __init__(self):
-        self.clear()
-
-    def clear(self):
-        self.state = State()
-        self.incoming = [State() for _ in range(len(Direction))]
-        self.outgoing = [State() for _ in range(len(Direction))]
-        self.reproduce = [False for _ in range(len(Direction))]
-
-class Cell(States):
-    # Resources
-    water = 0
-    energy = 0
-    # State
-    colour = None
-    wsat = 128.0
-    permeability = (1.0/8.0)
-    water_pressure_external = []
-    pressure_gradient = []
-    flux = []
-
-    def __init__(self):
-        super().__init__()
-        self.water_pressure_external = [0.0] * len(Direction)
-        self.pressure_gradient = [0.0] * len(Direction)
-        self.flux = [0] * len(Direction)
-
-    def update_water(self):
-        # Calculate the internal water pressure
-        if self.water < self.wsat:
-            pressure = UNSATURATED_PRESSURE_GRADIENT * self.water
-        else:
-            pressure = SATURATED_PRESSURE_GRADIENT * self.water
-
-        # Calculate the pressure gradient on each face
-        water_pressure = [
-            (pressure - self.water_pressure_external[direction]) * self.permeability
-            for direction in range(len(Direction))
-        ]
-        water_pressure[Direction.BELOW.value] += self.water * self.permeability
-
-        for direction in range(len(Direction)):
-            self.flux[direction] = (water_pressure[direction])
-
-    def update_flux(self):
-        new_flux = [0] * 6
-        total = 0
-        water_orig = self.water
-        while total < water_orig and max(self.flux) > 0:
-            pos = self.flux.index(max(self.flux))
-            if self.water > self.flux[pos]:
-                new_flux[pos] = self.flux[pos]
-                total += self.flux[pos]
-                self.water -= self.flux[pos]
-                self.flux[pos] = 0
-            else:
-                new_flux[pos] = self.water
-                total += self.water
-                self.flux[pos] -= self.water
-                self.water = 0
-        self.flux = new_flux
-
-    def apply_flux(self, incoming):
-        self.water += incoming
-
-    def update(self, grid):
-        pass
-
-class Air(Cell):
-    colour = (0.0, 0.0, 0.0, 0.0)
-
-    def __init__(self):
-        super().__init__()
-        self.water_pressure_external = [10000.0] * len(Direction)
-
-    def update_water(self):
-        pass
-
-    def update_flux(self):
-        pass
-
-    def update(self, grid):
-        pass
-
-    def apply_flux(self, incoming):
-        if incoming > 0:
-            print("ERROR: {}".format(incoming))
-            exit()
-        pass
-
-def interpolate(col1, col2, s):
-    return (
-        (col2[0] * s) + (col1[0] * (1 - s)),
-        (col2[1] * s) + (col1[1] * (1 - s)),
-        (col2[2] * s) + (col1[2] * (1 - s)),
-        #(col2[3] * s) + (col1[3] * (1 - s)),
-        s if s > 0.2 else 0.0
-    )
-
-class Soil(Cell):
-    colour = (0.0, 0.0, 0.0, 0.0)
-
-    def __init__(self):
-        super().__init__()
-        self.water = 0 #random.randint(1, 10)
-
-    def update(self, grid):
-        scale = min(self.water, 1.0) / 1.0
-        rock = (0.0, 0.0, 0.0, 0.8)
-        water = (0.075, 0.416, 0.936, 0.8)
-
-        self.colour = interpolate(water, water, scale)
-
-class Rock(Cell):
-    colour = (0.6, 0.6, 0.6, 1.0)
-
-    def __init__(self):
-        super().__init__()
-        self.water_pressure_external = [10000.0] * len(Direction)
-
-    def update_water(self):
-        pass
-
-    def update_flux(self):
-        pass
-
-    def update(self, grid):
-        pass
-
-    def apply_flux(self, incoming):
-        if incoming > 0:
-            print("ERROR: {}".format(incoming))
-            exit()
-        pass
 
 class Grid():
     width = 8
@@ -193,8 +50,6 @@ class Grid():
     # Threading
     render_lock = None
     colours = []
-
-
 
     def populate(self):
         self.grid = []
@@ -212,11 +67,15 @@ class Grid():
                         height = 0.0 + (2.0 * math.sin(0.13 * math.pi * xnorm) + math.cos(0.1 * math.pi * ynorm))
                         if znorm < height:
                             column.append(Rock())
-                        else:
+                        elif znorm < height + 2:
                             column.append(Soil())
+                        else:
+                            column.append(Air())
                 row.append(column)
             self.grid.append(row)
-        self.grid[3][3][7].water = 128
+        self.grid[3][1][4].water = 128
+
+        self.grid[3][3][5] = Plant()
 
         for x in range(self.width):
             for y in range(self.depth):
@@ -227,7 +86,6 @@ class Grid():
                             reverse = opposite(direction).value
                             neighbour = self.neighbour(x, y, z, direction)
                             neighbour.water_pressure_external[reverse] = 10000.0
-
 
         for x in range(self.width):
             row = []
@@ -312,9 +170,7 @@ class Grid():
                     cell.incoming[Direction.BEHIND.value] = self.cell(x, y + 1, z).outgoing[Direction.FRONT.value]
                     cell.incoming[Direction.BELOW.value] = self.cell(x, y, z - 1).outgoing[Direction.ABOVE.value]
                     cell.incoming[Direction.ABOVE.value] = self.cell(x, y, z + 1).outgoing[Direction.BELOW.value]
-                    
                     cell.update_water()
-
 
     def postupdate(self):
         # Transfer the pressures
@@ -383,8 +239,6 @@ class Grid():
                     if direction != None:
                         self.grid[x][y][z] = self.neightbour(x, y, z, direction).deepcopy()
 
-    count = 1
-
     def update(self):
         self.preupdate()
 
@@ -405,6 +259,7 @@ class Grid():
             print(line)
             
     def grid_update(self):
+        input("Press Enter to continue...")
         while True:
             self.update()
             with self.render_lock:
@@ -428,9 +283,9 @@ class Grid():
                         voxel = self.voxels[count]
                         colour = self.colours[x][y][z]
                         voxel.prop.color = "#{:02x}{:02x}{:02x}".format(
-                            int(colour[0] * 256),
-                            int(colour[1] * 256),
-                            int(colour[2] * 256)
+                            int(colour[0] * 255),
+                            int(colour[1] * 255),
+                            int(colour[2] * 255)
                         )
                         voxel.prop.opacity = colour[3]
                         count += 1
@@ -443,11 +298,9 @@ class Grid():
         # Create the scene
         pl, self.voxels = vxm.draw(self.width, self.depth, self.height)
         pl.add_callback(lambda : self.update_colours(), interval=1000)
-
         print("...Prepared")
+
         self.start_grid_thread()
-        print("...Prepared2")
-        input("Press Enter to continue...")
 
         while True:
             pl.render()
@@ -457,9 +310,10 @@ class Grid():
 if __name__ == "__main__":
     grid = Grid()
     grid.main()
-    
-    
-    
-    
+
+
+
+
+
